requirements.txt
# Core Django packages
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0
python-decouple==3.8

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter
django-redis==5.4.0  # Redis cache backend
dj-database-url==2.1.0  # Database URL parsing

# Image processing
Pillow==10.4.0

# Task queue and caching
celery==5.3.4
redis==5.0.1

# Django extensions
django-extensions==3.2.3
whitenoise==6.6.0

# Web server
gunicorn==21.2.0

# WebSocket support
channels==4.0.0
channels-redis==4.1.0

# API documentation
drf-spectacular==0.27.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Performance monitoring
psutil==5.9.6
websocket-client==1.6.4

# Additional packages for better compatibility
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2
setuptools>=65.5.0  # Required for pkg_resources
